Move LiteCertificate to certificate/lite.rs
