+++ linera-core/src/node.rs
@@ -13,8 +13,8 @@ use linera_base::{
     identifiers::{BlobId, ChainId},
 };
 use linera_chain::{
-    data_types::{BlockProposal, LiteCertificate, Origin},
-    types::{Certificate, HashedCertificateValue},
+    data_types::{BlockProposal, Origin},
+    types::{Certificate, HashedCertificateValue, LiteCertificate},
     ChainError,
 };
 use linera_execution::{
