+++ linera-core/src/value_cache.rs
@@ -12,10 +12,7 @@ use std::{any::type_name, sync::LazyLock};
 use std::{borrow::Cow, hash::Hash, num::NonZeroUsize};
 
 use linera_base::{crypto::CryptoHash, data_types::Blob, identifiers::BlobId};
-use linera_chain::{
-    data_types::LiteCertificate,
-    types::{Certificate, HashedCertificateValue},
-};
+use linera_chain::types::{Certificate, HashedCertificateValue, LiteCertificate};
 use lru::LruCache;
 use tokio::sync::Mutex;
 #[cfg(with_metrics)]
