+++ linera-rpc/src/lib.rs
@@ -27,7 +27,7 @@ pub use node_provider::NodeOptions;
 #[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
 #[cfg_attr(with_testing, derive(Eq, PartialEq))]
 pub struct HandleLiteCertRequest<'a> {
-    pub certificate: linera_chain::data_types::LiteCertificate<'a>,
+    pub certificate: linera_chain::types::LiteCertificate<'a>,
     pub wait_for_outgoing_messages: bool,
 }
